% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_candidates.R
\name{filter_candidates_ui}
\alias{filter_candidates_ui}
\alias{filter_candidates_server}
\title{filter structure and formula from mcn of ui}
\usage{
filter_candidates_ui(id)

filter_candidates_server(id, volumes, mcn_objects)
}
\description{
filter structure and formula from mcn of ui

filter structure and formula from mcn of server
}
