% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_classes.R
\name{filter_classes_ui}
\alias{filter_classes_ui}
\alias{filter_classes_server}
\title{filter classes from mcn of ui}
\usage{
filter_classes_ui(id)

filter_classes_server(id, volumes, mcn_objects)
}
\description{
filter classes from mcn of ui

filter classes from mcn of server
}
