% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_nebula.R
\name{visualize_nebula_ui}
\alias{visualize_nebula_ui}
\alias{visualize_nebula_server}
\title{create layouts for Parent-Nebula or Child-Nebulae visualizations.}
\usage{
visualize_nebula_ui(id)

visualize_nebula_server(id, volumes, mcn_objects)
}
\description{
create layouts for Parent-Nebula or Child-Nebulae visualizations.

create layouts for Parent-Nebula or Child-Nebulae visualizations.
}
