% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_MCnebulaApp}
\alias{run_MCnebulaApp}
\title{Run the Shiny Application}
\usage{
run_MCnebulaApp(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  maxRequestSize = 1000,
  ...
)
}
\description{
Run the Shiny Application
}
